% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caspex_chipatlas.R
\name{download_chipatlas_experiment_list}
\alias{download_chipatlas_experiment_list}
\title{Download (if missing) and cache ChIP-Atlas experimentList.tab}
\usage{
download_chipatlas_experiment_list(force = FALSE, quiet = FALSE)
}
\arguments{
\item{force}{logical; re-download even if cached.}

\item{quiet}{passed to download.file().}
}
\value{
path to the cached TSV.
}
\description{
Download (if missing) and cache ChIP-Atlas experimentList.tab
}
\examples{
\dontrun{
exp <- download_chipatlas_experiment_list()
}
}
