% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caspex_analysis.R
\name{plot_grna_positions}
\alias{plot_grna_positions}
\title{Plot 1 — gRNA positions ruler}
\usage{
plot_grna_positions(pos_map, gene_info, upstream, downstream)
}
\arguments{
\item{pos_map}{(see function body).}

\item{gene_info}{(see function body).}

\item{upstream}{(see function body).}

\item{downstream}{(see function body).}
}
\description{
Plot 1 — gRNA positions ruler
}
