% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caspex_extras.R
\name{plot_volcano_per_region}
\alias{plot_volcano_per_region}
\title{One volcano per region, TFs highlighted.}
\usage{
plot_volcano_per_region(
  result,
  pval_thresh = 0.05,
  lfc_thresh = 1,
  label_top = 8
)
}
\arguments{
\item{result}{(see function body).}

\item{pval_thresh}{(see function body).}

\item{lfc_thresh}{(see function body).}

\item{label_top}{(see function body).}
}
\description{
One volcano per region, TFs highlighted.
}
