% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caspex_extras.R
\name{plot_event_jackknife}
\alias{plot_event_jackknife}
\title{Plot results of run_event_jackknife()}
\usage{
plot_event_jackknife(jk_result, top_n = 40)
}
\arguments{
\item{jk_result}{(see function body).}

\item{top_n}{(see function body).}
}
\value{
A ggplot.
}
\description{
Visualises which TF events survive after dropping each region in turn.
}
