% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caspex_extras.R
\name{plot_region_correlation}
\alias{plot_region_correlation}
\title{Pearson correlation heatmap of region logFC profiles.}
\usage{
plot_region_correlation(result, method = "pearson")
}
\arguments{
\item{result}{(see function body).}

\item{method}{(see function body).}
}
\description{
Pearson correlation heatmap of region logFC profiles.
}
