% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caspex_analysis.R
\name{rescan_motifs}
\alias{rescan_motifs}
\title{Re-run motif scan with a different TF set or threshold}
\usage{
rescan_motifs(result, tf_names, threshold_frac = 0.8)
}
\arguments{
\item{result}{(see function body).}

\item{tf_names}{(see function body).}

\item{threshold_frac}{(see function body).}
}
\description{
Re-run motif scan with a different TF set or threshold
}
