% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caspex_chipatlas.R
\name{run_chipatlas_scan}
\alias{run_chipatlas_scan}
\title{Batch fetch for a set of TFs}
\usage{
run_chipatlas_scan(
  tfs,
  gene_info,
  promoter_info,
  upstream = 2500,
  downstream = 500,
  threshold = "05",
  max_experiments = 50,
  special_interest_gene = NULL,
  special_interest_cap = NULL,
  quiet = TRUE
)
}
\arguments{
\item{tfs}{(see function body).}

\item{gene_info}{(see function body).}

\item{promoter_info}{(see function body).}

\item{upstream}{(see function body).}

\item{downstream}{(see function body).}

\item{threshold}{(see function body).}

\item{max_experiments}{(see function body).}

\item{special_interest_gene}{(see function body).}

\item{special_interest_cap}{(see function body).}

\item{quiet}{(see function body).}
}
\value{
named list tf -> data.frame (or NULL if no peaks in window).
}
\description{
Batch fetch for a set of TFs
}
